working on delete and update functionality

+++ src/main/java/com/codeup/springblog/repos/PostRepository.java
@@ -7,8 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 public interface PostRepository extends JpaRepository<Post, Long> {
 
     @Query("from Post a where a.id = ?1")
-    Post getPostById(int id);
+    Post getPostById(Long id);
 
     Post getByBody(String body);
+
+
 
 }
